+++ src/mkvtoolnix-gui/merge/track.cpp
@@ -256,7 +256,6 @@ Track::saveSettings(Util::ConfigFile &settings)
   settings.setValue("defaultTrackFlag",              m_defaultTrackFlag);
   settings.setValue("forcedTrackFlag",               m_forcedTrackFlag);
   settings.setValue("stereoscopy",                   m_stereoscopy);
-  settings.setValue("naluSizeLength",                m_naluSizeLength);
   settings.setValue("cues",                          m_cues);
   settings.setValue("aacIsSBR",                      m_aacIsSBR);
   settings.setValue("reduceAudioToCore",             m_reduceAudioToCore);
@@ -312,7 +311,6 @@ Track::loadSettings(MuxConfig::Loader &l) {
   m_defaultTrackFlag              = l.settings.value("defaultTrackFlag").toInt();
   m_forcedTrackFlag               = l.settings.value("forcedTrackFlag").toInt();
   m_stereoscopy                   = l.settings.value("stereoscopy").toInt();
-  m_naluSizeLength                = l.settings.value("naluSizeLength").toInt();
   m_cues                          = l.settings.value("cues").toInt();
   m_aacIsSBR                      = l.settings.value("aacIsSBR").toInt();
   m_reduceAudioToCore             = l.settings.value("reduceAudioToCore").toBool();
@@ -445,9 +443,6 @@ Track::buildMkvmergeOptions(MkvmergeOptionBuilder &opt)
 
     if (m_stereoscopy)
       opt.options << Q("--stereo-mode") << Q("%1:%2").arg(sid).arg(m_stereoscopy - 1);
-
-    if (m_naluSizeLength)
-      opt.options << Q("--nalu-size-length") << Q("%1:%2").arg(sid).arg(m_naluSizeLength);
 
     if (m_compression != TrackCompression::Default)
       opt.options << Q("--compression") << Q("%1:%2").arg(sid).arg(TrackCompression::None == m_compression ? Q("none") : Q("zlib"));