+++ src/output/p_hevc.cpp
@@ -25,7 +25,7 @@
 
 class hevc_video_packetizer_private_c {
 public:
-  int nalu_size_len_src{}, nalu_size_len_dst{};
+  int nalu_size_len{};
   int64_t max_nalu_size{};
   bool rederive_timestamp_order{};
   std::unique_ptr<mtx::hevc::es_parser_c> parser;
@@ -42,7 +42,8 @@ hevc_video_packetizer_c(generic_reader_c *p_reader,
 {
   m_relaxed_timestamp_checking = true;
 
-  setup_nalu_size_len_change();
+  if (23 <= m_ti.m_private_data->get_size())
+    p_ptr->nalu_size_len = (m_ti.m_private_data->get_buffer()[21] & 0x03) + 1;
 
   set_codec_private(m_ti.m_private_data);
 }
@@ -95,9 +96,6 @@ hevc_video_packetizer_c::process(packet_cptr packet) {
 
   m_ref_timestamp = packet->timestamp;
 
-  if (p.nalu_size_len_dst && (p.nalu_size_len_dst != p.nalu_size_len_src))
-    change_nalu_size_len(packet);
-
   add_packet(packet);
 
   return FILE_STATUS_MOREDATA;
@@ -116,90 +114,6 @@ hevc_video_packetizer_c::can_connect_to(generic_packetizer_c *src,
   }
 
   return CAN_CONNECT_YES;
-}
-
-void
-hevc_video_packetizer_c::setup_nalu_size_len_change() {
-  if (!m_ti.m_private_data || (23 > m_ti.m_private_data->get_size()))
-    return;
-
-  auto &p             = *p_func();
-  auto private_data   = m_ti.m_private_data->get_buffer();
-  auto const len_pos  = 21;
-  p.nalu_size_len_src = (private_data[len_pos] & 0x03) + 1;
-  p.nalu_size_len_dst = p.nalu_size_len_src;
-
-  if (!m_ti.m_nalu_size_length || (m_ti.m_nalu_size_length == p.nalu_size_len_src))
-    return;
-
-  p.nalu_size_len_dst   = m_ti.m_nalu_size_length;
-  private_data[len_pos] = (private_data[len_pos] & 0xfc) | (p.nalu_size_len_dst - 1);
-  p.max_nalu_size       = 1ll << (8 * p.nalu_size_len_dst);
-
-  set_codec_private(m_ti.m_private_data);
-}
-
-void
-hevc_video_packetizer_c::change_nalu_size_len(packet_cptr const &packet) {
-  unsigned char *src = packet->data->get_buffer();
-  int size           = packet->data->get_size();
-
-  if (!src || !size)
-    return;
-
-  std::vector<int> nalu_sizes;
-
-  auto &p     = *p_func();
-  int src_pos = 0;
-
-  // Find all NALU sizes in this packet.
-  while (src_pos < size) {
-    if ((size - src_pos) < p.nalu_size_len_src)
-      break;
-
-    int nalu_size = 0;
-    int i;
-    for (i = 0; i < p.nalu_size_len_src; ++i)
-      nalu_size = (nalu_size << 8) + src[src_pos + i];
-
-    if ((size - src_pos - p.nalu_size_len_src) < nalu_size)
-      nalu_size = size - src_pos - p.nalu_size_len_src;
-
-    if (nalu_size > p.max_nalu_size)
-      mxerror_tid(m_ti.m_fname, m_ti.m_id, fmt::format(Y("The chosen NALU size length of {0} is too small. Try using '4'.\n"), p.nalu_size_len_dst));
-
-    src_pos += p.nalu_size_len_src + nalu_size;
-
-    nalu_sizes.push_back(nalu_size);
-  }
-
-  // Allocate memory if the new NALU size length is greater
-  // than the previous one. Otherwise reuse the existing memory.
-  if (p.nalu_size_len_dst > p.nalu_size_len_src) {
-    int new_size = size + nalu_sizes.size() * (p.nalu_size_len_dst - p.nalu_size_len_src);
-    packet->data = memory_c::alloc(new_size);
-  }
-
-  // Copy the NALUs and write the new sized length field.
-  unsigned char *dst = packet->data->get_buffer();
-  src_pos            = 0;
-  int dst_pos        = 0;
-
-  size_t i;
-  for (i = 0; nalu_sizes.size() > i; ++i) {
-    int nalu_size = nalu_sizes[i];
-
-    int shift;
-    for (shift = 0; shift < p.nalu_size_len_dst; ++shift)
-      dst[dst_pos + shift] = (nalu_size >> (8 * (p.nalu_size_len_dst - 1 - shift))) & 0xff;
-
-    memmove(&dst[dst_pos + p.nalu_size_len_dst], &src[src_pos + p.nalu_size_len_src], nalu_size);
-
-    src_pos += p.nalu_size_len_src + nalu_size;
-    dst_pos += p.nalu_size_len_dst + nalu_size;
-  }
-
-  packet->data->set_size(dst_pos);
 }
 
 void
@@ -226,7 +140,7 @@ hevc_video_packetizer_c::process_rederiving_timestamp_order(packet_t &packet) {
   if (packet.has_timestamp())
     p.parser->add_timestamp(packet.timestamp);
 
-  p.parser->add_bytes_framed(packet.data, p.nalu_size_len_src);
+  p.parser->add_bytes_framed(packet.data, p.nalu_size_len);
 
   flush_frames();
 }