+++ src/output/p_hevc.h
@@ -40,8 +40,6 @@ public:
 
 protected:
   virtual void extract_aspect_ratio();
-  virtual void setup_nalu_size_len_change();
-  virtual void change_nalu_size_len(packet_cptr const &packet);
 
   virtual void process_rederiving_timestamp_order(packet_t &packet);